tests/xspec-show-modal-focusing-steps.html

<!DOCTYPE html>
<html>
<head>
<meta charset=utf-8>
<script src="./dist/index.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/common.js"></script>
<script>
promise_test(function() {
  return waitUntilLoadedAndAutofocused().then(function() {
        outerButton = document.getElementById('outer-button');
        // 2021-03-31 - Manually focused until Webkit autofocus behaviour is
        // resolved: https://bugs.webkit.org/show_bug.cgi?id=203139
        if (document.activeElement !== outerButton) {
            outerButton.focus();
        }
        assert_equals(document.activeElement, outerButton);

        // Test that focus goes to body if the dialog has no focusable elements, including itself
        var outerDialog = document.getElementById('outer-dialog');
        outerDialog.showModal();
        assert_equals(document.activeElement, outerDialog);

        // Test that an autofocus element in the dialog gets focus.
        var dialog = document.getElementById('dialog');
        dialog.showModal();
        autofocusButton = document.getElementById('autofocus-button');
        assert_equals(document.activeElement, autofocusButton);
        dialog.close();

        // Focusable elements without autofocus should not gain focus
        autofocusButton.parentNode.removeChild(autofocusButton);
        dialog.showModal();
        assert_equals(document.activeElement, dialog);
        dialog.close();

        // ... or else the dialog itself gets focus.;
        var buttons = dialog.querySelectorAll('button');
        for (var i = 0; i < buttons.length; ++i)
            buttons[i].hidden = true;
        dialog.showModal();
        assert_equals(document.activeElement, dialog);
        dialog.close();

        document.getElementById('outer-dialog').close();
  });
}, "focus when a modal dialog is opened");
</script>
</head>
<body>
<button id="outer-button" autofocus></button>
<dialog id="outer-dialog">
    <dialog id="dialog" tabindex=0>
        <button disabled></button>
        <dialog>
            <button autofocus></button>
        </dialog>
        <button id="first-button"></button>
        <div>
            <span>
                <button id="autofocus-button" autofocus></button>
            </span>
        </div>
        <button id="final-button"></button>
    </dialog>
</dialog>
</body>
</html>
